6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long int c;
 	long long int d;
 	float f;
-printf("Size of a char: %lu byte(s)\n", sizeof(a));
-printf("Size of an int: %lu bytes(s)\n", sizeof(b));
-printf("size of a long int: %lu bytes(s)\n", sizeof(c));
-printf("size of a long long int: %lu bytes(s)\n", sizeof(d));
-printf("size of a float: %lu bytes(s)\n", sizeof(f));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
+printf("Size of an int: %lu bytes(s)\n", (unsigned long)sizeof(b));
+printf("size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(c));
+printf("size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(d));
+printf("size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
